+++ src/main/java/com/rudolph/Weevo/notification/domain/enums/NotiType.java
@@ -1,4 +1,4 @@
-package com.rudolph.Weevo.notification.enums;
+package com.rudolph.Weevo.notification.domain.enums;
 
 public enum NotiType {
     LECTURE_MATCHED,
